refactor: move gensuperdrop into seperate file
