+++ libs/runcleo/CMakeLists.txt
@@ -14,6 +14,7 @@ set(LIBRARY_OUTPUT_PATH ${CMAKE_BINARY_DIR}/lib)
 set(SOURCES
 "creategbxs.cpp"
 "createsupers.cpp"
+"gensuperdrop.cpp"
 )
 # must use STATIC (not(!) SHARED) lib for linking to executable if build is CUDA enabled with Kokkos
 add_library("${LIBNAME}" STATIC ${SOURCES})
