+++ libs/runcleo/createsupers.hpp
@@ -23,101 +23,13 @@
 #define LIBS_RUNCLEO_CREATESUPERS_HPP_
 
 #include <Kokkos_Core.hpp>
-#include <array>
 #include <iostream>
-#include <memory>
 #include <stdexcept>
 #include <string>
 
 #include "../kokkosaliases.hpp"
 #include "gridboxes/sortsupers.hpp"
-#include "initialise/initialconditions.hpp"
-#include "superdrops/superdrop.hpp"
-
-/**
- * @brief Struct that holds data for the initial conditions of super-droplets.
- *
- * This struct holds data for the initial conditions of some properties of super-droplets
- * and provides an operator which returns a super-droplet generated using them.
- */
-class GenSuperdrop {
- private:
-  unsigned int nspacedims; /**< Number of spatial dimensions. */
-  std::shared_ptr<Superdrop::IDType::Gen> sdIdGen;
-  /**< Pointer Superdrop::IDType object for super-droplet ID generation. */
-  InitSupersData initdata; /**< instance of InitSupersData for initialising superdrops. */
-
-  /**
-   * @brief Returns initial spatial coordinates of the kk'th super-droplet.
-   *
-   * A coordinate is only copied from the corresponding coords vector if that coordinate is
-   * consistent with the number of spatial dimensions of the model. Otherwise, the coordinate is set
-   * to 0.0. For example, if the model is 1-D, only coord3 is obtained from the initial data vector;
-   * coord1 and coord2 are set to 0.0.
-   *
-   * @param kk The index of the super-droplet in the initial
-   * data (0 <= kk < total number of super-droplets).
-   * @return An array containing the kk'th superdroplet's spatial
-   * coordinates (coord3, coord1, coord2).
-   */
-  std::array<double, 3> coords_at(const unsigned int kk) const;
-
-  /* helper function to return a superdroplet's attributes
-  at position kk in the initial conditions data. All
-  superdroplets created with same solute properties */
-  /**
-   * @brief Function returns a superdroplet's attributes
-   * from position 'kk' in the initial conditions data. All
-   * super-droplets have the same solute properties.
-   *
-   * @param kk The index of the super-droplet in the initial
-   * data (0 <= kk < total number of super-droplets).
-   * @return The attributes of the superdrop from index 'kk'.
-   */
-  SuperdropAttrs attrs_at(const unsigned int kk) const;
-
- public:
-  /**
-   * @brief Constructs a GenSuperdrop instance.
-   *
-   * This constructor initializes a GenSuperdrop instance using the provided
-   * SuperdropInitConds instance to fetch initial data.
-   *
-   * @tparam SuperdropInitConds The type of the data for super-droplets' initial conditions.
-   * @param sdic The instance of the data for super-droplets' initial conditions.
-   */
-  template <typename SuperdropInitConds>
-  explicit GenSuperdrop(const SuperdropInitConds &sdic)
-      : nspacedims(sdic.get_nspacedims()), sdIdGen(std::make_shared<Superdrop::IDType::Gen>()) {
-    sdic.fetch_data(initdata);
-  }
-
-  /**
-   * @brief Generate a super-droplet using initial data for the kk'th superdrop.
-   *
-   * This function returns a superdrop generated from the specified position
-   * 'kk' in the initial conditions data.
-   *
-   * @param kk The index of the superdrop to generate.
-   * @return The generated super-droplet.
-   */
-  Superdrop operator()(const unsigned int kk) const;
-};
-
-/**
- * @brief Create a view of super-droplets in (device) memory.
- *
- * This function creates an ordered view of superdrops in device memory, where the number
- * of superdrops is specified by the parameter `totnsupers`. The superdrops are
- * ordered by the gridbox indexes and generated using a generator which uses
- * the initial conditions referenced by the `SuperdropInitConds` type.
- *
- * @tparam SuperdropInitConds The type of the super-droplets' initial conditions data.
- * @param sdic The instance of the super-droplets' initial conditions data.
- * @return A view of super-droplets in device memory.
- */
-template <typename SuperdropInitConds>
-viewd_supers create_supers(const SuperdropInitConds &sdic);
+#include "runcleo/gensuperdrop.hpp"
 
 /**
  * @brief Return an initialised view of superdrops in device memory.
@@ -188,7 +100,7 @@ void print_supers(const viewd_constsupers supers);
  * This function creates an ordered view of superdrops in device memory, where the number
  * of superdrops is specified by the parameter `totnsupers`. The superdrops are
  * ordered by the gridbox indexes and generated using a generator which uses
- * the initial conditions referenced by the `SuperdropInitConds` type.
+ * the initial conditions provided by the `SuperdropInitConds` type.
  *
  * @tparam SuperdropInitConds The type of the super-droplets' initial conditions data.
  * @param sdic The instance of the super-droplets' initial conditions data.
@@ -277,4 +189,5 @@ viewd_supers::HostMirror initialise_supers_on_host(const SuperdropInitConds &sdi
 
   return h_supers;
 }
+
 #endif  // LIBS_RUNCLEO_CREATESUPERS_HPP_
